+++ editor/src/components/canvas/canvas-strategies/canvas-strategies.tsx
@@ -63,7 +63,7 @@ import type { InsertionSubject, InsertionSubjectWrapper } from '../../editor/edi
 import { generateUidWithExistingComponents } from '../../../core/model/element-template-utils'
 import { retargetStrategyToChildrenOfFragmentLikeElements } from './strategies/fragment-like-helpers'
 import { MetadataUtils } from '../../../core/model/element-metadata-utils'
-import { rearrangeGridStrategy } from './strategies/rearrangeGridStrategy'
+import { rearrangeGridStrategy } from './strategies/rearrange-grid-strategy'
 
 export type CanvasStrategyFactory = (
   canvasState: InteractionCanvasState,
